requirements.txt
requests==2.32.3          
python-dotenv==1.0.1      
pandas==2.2.3             
numpy==1.26.4             
tqdm==4.66.5             

py2neo==2021.2.4
neo4j>=5.15.0
networkx==3.3             
pyvis==0.3.2              
rdflib==7.0.0             

scikit-learn==1.5.2       
torch==2.4.1              
transformers==4.44.2      


fastapi==0.115.0          
uvicorn==0.30.6           

loguru==0.7.2             
beautifulsoup4==4.12.3    
openpyxl==3.1.5           

lxml
networkx
pyvis
matplotlib
streamlit
ollama
psycopg2-binary
Flask==3.0.0
pyjwt==2.8.0
cryptography==41.0.7
openai
google-generativeai